==> nbs/00_leaderboard.ipynb <==
# default_exp leaderboard

# Leaderboard

> Visualising the public leaderboard.

%load_ext autoreload
%autoreload 2

#export
import pandas as pd
from pathlib import Path
import os
import plotly.graph_objects as go
import plotly.express as px
from ashrae import inspection

pd.options.plotting.backend = "plotly"

#export
data_path = Path("../data")

data_path.ls()

csv_names = list(inspection.CSV_NAMES)
csv_names.append('ashrae')
csvs = inspection.get_csvs(data_path, csv_names=csv_names)
csvs

leaderboard = pd.read_csv(csvs['ashrae'], parse_dates=['SubmissionDate'])
leaderboard.head()

leaderboard['TeamId'].nunique(), len(leaderboard)

pareto = (leaderboard.sort_values('Score')
          .groupby('Score')
          .size()
          .rename('count')
          .reset_index()
          .assign(**{'cumulative %': lambda x: 100*x['count'].cumsum()/x['count'].sum()})
         )
pareto

px.line(pareto, x='Score', y='cumulative %')

leaderboard['Score'].plot(kind='hist')

leaderboard['Score'].describe(percentiles=[.05, .1, .25, .5, .75, .95])

leaderboard.plot(kind='scatter', x='SubmissionDate', y='Score')

leaderboard.sort_values('Score').head()
